Added component tests for ensembler
